app/requirements.txt
fastapi
uvicorn
openai
weaviate-client
langchain
langchain-community
langchain-weaviate
sentence-transformers
pandas
numpy